Refactoring views in authentication app into their own files
Added a views folder containing a view file for each logical group of views.
