+++ authentication/urls.py
@@ -1,28 +1,32 @@
 from django.urls import path
-from . import views
+from .views import (
+    test_views,
+    authentication_views,
+    user_views,
+)
 
 app_name = 'authentication'
 
 urlpatterns = [
     path(
         'connection-test/',
-        views.TestConnection.as_view(),
+        test_views.TestConnection.as_view(),
         name='test_connection'
     ),
     path(
-        'login-initial/', views.LoginInitial.as_view(),
+        'login-initial/', authentication_views.LoginInitial.as_view(),
         name='login-initial'
     ),
     path(
-        'login-final/', views.LoginFinal.as_view(),
+        'login-final/', authentication_views.LoginFinal.as_view(),
         name='login-final'
     ),
     path(
-        'user/', views.CreateUser.as_view(),
+        'user/', user_views.CreateUser.as_view(),
         name='create_user'
     ),
     path(
-        'user/', views.UpdateUser.as_view(),
+        'user/update/', user_views.UpdateUser.as_view(),
         name='update_user'
     ),
 ]
